Fix .ssh/config file permissions when cloning
Even if the file permissions ar `600` when git cloned it will be set to `644`.
Ssh does not like `644`.
Even though I'm changing the mode to `600` now there are no changed in `home-private` repo.
That's fine. I'm not going to do there anything manually so this should work 100% of the time.
Case closed.